LINE following added, also live video streaming added

+++ sw/airborne/modules/computer_vision/combine/inter_thread_data_IMAV2015.h
@@ -43,8 +43,8 @@ struct opticflow_result_t {
   float vel_x;            ///< The velocity in the x direction
   float vel_y;            ///< The velocity in the y direction
   
-  float deviation_x;      ///< Deviation of centroid from the center in x direction [pixel]
-  float deviation_y;      ///< Deviation of centroid from the center in y direction [pixel]
+  int16_t deviation_x;      ///< Deviation of centroid from the center in x direction [pixel]
+  int16_t deviation_y;      ///< Deviation of centroid from the center in y direction [pixel]
   uint16_t inlier;            ///< The number of inliers
 };
